Update 02a_cr_create_population.do
for job-running to run smoothly (cannot have same dataset name). This do-file will run several times for different outcomes, but this dataset exported is the same for all outcomes, so initially i didn't specify outcome in the dataset name. But it would cause job-running stop due to same dataset name